# docs/apidog.openapi.compat.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MLE Home Backend API",
    "version": "0.1.0",
    "description": "OpenAPI import file for Apidog. This describes the Dockerized local API for student linking, diary records, evidence uploads, OCR inference, privacy decisions, tutor summaries, reports, curriculum standards, and integration metadata."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Docker Compose API"
    }
  ],
...
  "paths": {
...
    "/v1/reports/students/{studentId}/stats": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Generate tutor statistics report",
        "operationId": "generateTutorStatsReport",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountIdHeader"
          },
          {
            "$ref": "#/components/parameters/AccountRoleHeader"
          },
          {
            "$ref": "#/components/parameters/StudentIdPath"
          },
          {
            "$ref": "#/components/parameters/FromQuery"
          },
          {
            "$ref": "#/components/parameters/ToQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Tutor statistics report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorStatsReportResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports/students/{studentId}/education-diary": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Generate guardian education diary report",
        "operationId": "generateEducationDiaryReport",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountIdHeader"
          },
          {
            "$ref": "#/components/parameters/AccountRoleHeader"
          },
          {
            "$ref": "#/components/parameters/StudentIdPath"
          },
          {
            "$ref": "#/components/parameters/FromQuery"
          },
          {
            "$ref": "#/components/parameters/ToQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          },
          {
            "name": "includeImages",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Only includes image object keys when evidence has been approved for image inclusion."
          }
        ],
        "responses": {
          "200": {
            "description": "Guardian education diary report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EducationDiaryReportResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "AccountIdHeader": {
        "name": "x-account-id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "example": "parent_demo"
        }
      },
      "AccountRoleHeader": {
        "name": "x-account-role",
        "in": "header",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/AccountRole"
        }
      },
      "StudentIdPath": {
        "name": "studentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "example": "student_demo"
        }
      },
...
      "FromQuery": {
        "name": "from",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "ToQuery": {
        "name": "to",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 50
        }
      }
    },
...
    "schemas": {
      "AccountRole": {
        "type": "string",
        "enum": [
          "parent",
          "child",
          "tutor",
          "integration"
        ],
        "example": "parent"
      },
      "Subject": {
        "type": "string",
        "enum": [
          "maths",
          "english",
          "science",
          "pe",
          "art",
          "humanities",
          "life-skills",
          "other"
        ]
      },
...
      "CurriculumTag": {
        "type": "object",
        "required": [
          "subject"
        ],
        "properties": {
          "standardId": {
            "type": "string"
          },
          "keyStage": {
            "type": "string",
            "example": "KS2"
          },
          "subject": {
            "type": "string",
            "example": "maths"
          },
          "topic": {
            "type": "string",
            "example": "Fractions"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "example": 0.8
          }
        }
      },
      "EvidenceReference": {
        "type": "object",
        "required": [
          "evidenceId",
          "includeInGuardianReport",
          "includeImageInReports"
        ],
        "properties": {
          "evidenceId": {
            "type": "string"
          },
          "objectKey": {
            "type": "string"
          },
          "mimeType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "description": {
            "type": "string"
          },
          "includeInGuardianReport": {
            "type": "boolean",
            "default": true
          },
          "includeImageInReports": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "DiaryEntry": {
        "type": "object",
        "required": [
          "_id",
          "studentId",
          "occurredAt",
          "durationMinutes",
          "subject",
          "title",
          "createdByAccountId",
          "createdByRole",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "example": "entry_5f3029c9-6c9e-4432-aab2-8a47209f1ce2"
          },
          "studentId": {
            "type": "string",
            "example": "student_demo"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationMinutes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1440,
            "example": 60
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "example": "Fractions practice"
          },
          "description": {
            "type": "string",
            "maxLength": 4000
          },
          "location": {
            "type": "string",
            "maxLength": 240
          },
          "tutorAccountId": {
            "type": "string"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            }
          },
          "curriculumTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurriculumTag"
            }
          },
          "shareWithTutorIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdByAccountId": {
            "type": "string",
            "example": "parent_demo"
          },
          "createdByRole": {
            "$ref": "#/components/schemas/AccountRole"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SubjectBreakdown": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "example": "maths"
          },
          "minutes": {
            "type": "integer",
            "example": 120
          },
          "hours": {
            "type": "number",
            "example": 2
          },
          "entries": {
            "type": "integer",
            "example": 2
          }
        }
      },
      "DiaryStats": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "example": "student_demo"
          },
          "from": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          },
          "totalMinutes": {
            "type": "integer",
            "example": 180
          },
          "totalHours": {
            "type": "number",
            "example": 3
          },
          "entryCount": {
            "type": "integer",
            "example": 3
          },
          "externalTutorMinutes": {
            "type": "integer",
            "example": 60
          },
          "subjectBreakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectBreakdown"
            }
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TutorStatsReportResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "reportType": {
                "type": "string",
                "example": "tutor-stats"
              },
              "access": {
                "type": "string",
                "example": "statistics-only"
              },
              "generatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "studentId": {
                "type": "string"
              },
              "stats": {
                "$ref": "#/components/schemas/DiaryStats"
              }
            }
          }
        }
      },
      "EducationDiaryReportResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "reportType": {
                "type": "string",
                "example": "guardian-education-diary"
              },
              "access": {
                "type": "string",
                "example": "guardian-controlled-evidence"
              },
              "generatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "studentId": {
                "type": "string"
              },
              "entries": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DiaryEntry"
                }
              }
            }
          }
        }
      },
...
    }
  }
}